services/ai-proctor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
mediapipe==0.10.14
Pillow==11.1.0
numpy==1.26.4
